:wrench: chore: Update environment variables and configurations
This commit updates the `.env.example` file with the latest environment variables and configurations. It also updates the `next.config.mjs` file to include validation of environment variables . Additionally, the `env.mjs` file is modified to reflect the changes in the environment variable names. These updates ensure that the application is using the correct environment variables and configurations for local development.