feat: Add 'status' prop to MenuProps type
This commit adds a new 'status' prop to the MenuProps type in the 'types.ts' file. The 'status' prop is of type string and is used to pass the status information to the Menu component. This change improves the type safety and provides a clear definition of the props required by the Menu component.