feat: Add gtag tracking functionality
This commit adds the `gtag.ts` file which contains the necessary functions for tracking page views and events using Google Analytics. The `pageview` function is used to track page views by sending the `url` as a parameter to the `window.gtag` function. The `event` function is used to track events by sending the `action`, `category`, `label`, and `value` as parameters to the `window.gtag` function. This addition enhances the tracking capabilities of the application.